backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
starlette>=0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
# psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9

# Email
sendgrid==6.11.0

# Media & Files
cloudinary==1.38.0
aiofiles==23.2.1
Pillow==10.2.0
boto3==1.34.69

# WebSocket & WebRTC
websockets==12.0
aiortc==1.6.0

# Redis (Caching)
redis==5.0.1
hiredis==2.3.2

# Environment & Configuration
python-dotenv==1.0.0

# Two-Factor Authentication (Future)
pyotp==2.9.0

# Utilities
python-dateutil==2.8.2
python-json-logger==2.0.7
